Only update _Contente-Type_ if the original was **SOAP**
By updating this, we can be sure that the _Content-Type_ is only set
once from SOAP to MIME and not every time a new `Attachment` is added to
the `AS4Message`.
By changing this, we don't have to manually set the _Content-Type_
afterwards in the `MimeMessageSerializer` because this is already
checked in the `AddAttachment()` call.

 #8285