Update download Message Body with Primary Key instead of EBMS Id
Because we don't use the EBMS Id as "unique" identifier (and so, have
multiple entries with the same value), we need another way of
identifying these messages.

By updating the `MonitorService` in the Frontend API and the
`message.component.html` files, we're sure that we send the right value
(Primary Key = `long`) and we retrieve it the right way (filtering on
the Primary Key = `long`).

 #8604