Update Dynamic Discovery information in the Sending PMode
The backend has changed the Dynamic Discovery functionality in such a
way that we only need to specify which "profile" implementation we
should use for the retrieval of the SMP Configuration.
Each profile implementation has its own properties to configure itself;
which differ from each profile implementation.

This commit contains:
- Load the `IDynamicDiscoveryProfile` implementations in the backend
- Expose these implementations with the route:
`api/runtime/getdynamicdiscoveryprofiles`
- Update the `getall` call and model in the frontend to also include these
profiles
- Update the Sending PMode UI with a combobox to select the profile and
a dynamic table to insert the custom settings

 #8405