Use the default Transformers from the BE in the UI
As enhancement of the UI, we should only show the `ITransformer`
implementation that are relevant to the corresponding `Agent` the user
want to create.
The Backend already has a API for this called the
`AgentProvider.GetDefaultTransformerForAgentType()` static method that's
exposed via the `/api/configuration/getdefaulttransformer` API call.

This commit contains:
- The verification that the `TransformerConfigEntry` model is
serializable
- The agent component in the UI to only show the needed transformers
- Update the `Transformer` type to correspond with the new response type
of the API
- Update the `ConfigurationController` to make sure that the returned
model of the Backend is mapped to a `ItemType` that we can easily plug
in the already implemented Frontend components

 #8610