Make transformers configurable in agent UI
Before, `ITransformer` implementation hasn't got any custom settings to
configure; now with the introduction of the Static Submit agent we do
have some settings that must be configured at runtime.

This commit contains the creation of a custom `transformer.component.ts`
to introduce not only a combobox for the type but also a dynamic table
that gets populated for the selected transformer type.

 #8404