PMode Id and Filname alteration
Before, when renaming a PMode via the UI; we only renamed the filename
of the PMode (not the Id itself).

This commit alters this behavior that it alters the Id (and the
filename) of the PMode. When the Id of the PMode contains invalid
filename charachers; these chars will be filtered out.

The Frontend needs to initialize the `Config` instance itself when it's
runned seperataly; so the `Program` is also allterd.

 #8701
 #8692
 #8615