Add Component Test for the Outbound Processing Agent
This newly created test exposed the problem with the serialization of
the `AS4Message` when the message is being processed by the Compression
step inside the Outbound Processing Agent.

Since this step will set the `as4message.EnvelopeDocument` back to
`null`; the property `IsMultihopMessage` is reset to `false` even when
the original was indeed a multihop message.

This commit changes the property inside the `AS4Message` to set a single
time the `IsMultihopMessage` inside the _Factory Method_ for the
`AS4Message` where we pass in the `XmlDocument` (envelope document).

 #8758