Use everywhere the Primary Key instead of the EbmsMessageId
This commit contains the necessary changes to use the Primary Key
instead of the EbmsMessageId in the Frontend.

There is also a change in the `MonitorController` that now explictly
uses a `long` instead of a `long` serialized as a `string`.

The `relatedmessages.component.html` also has a change that maybe should
be in the v3.1 version

 #8774