Upgrade the 'ReceiveMessageTransformer' to use in a Static Receive
Contains the upgrade of the existing `ReceiveMessageTransformer` to use
in a _Static Receive_ scenario.
This means that this transformer is now able to assign a default
_Receiving PMode_ to the created `MessagingContext` if there's a setting
that links to an existing Receiving PMode.

Since this setting is an optinal one; we don't use the
`_config.GetReceivingPMode()` call because this would mean that we are
sure that there's is a PMode.
If it occurs that the setting links to a non-existing PMode, or is not
configured proparly; we log a warning stating that there's maybe
something wrong with the settings for this transformer.

 #8959