Updates the Retry Reliability with default values
The retry count has now a default value of `4` and the retry intervalis
default set to an interval of a single day. (both defaults used from the
Logic Apps defaults for error handling).

This change of defaults has also effect on the validation tests for the
sending and receiving pmodes (and their schema's). This commit contains
also the necessary changes to show these defaults in the UI (accompanied
with UI tests).
 #9245
 #9247