Updates the staging script to also include the 'Ensure.That.dll'
This library is not only used in the tests of the Frontend, but also in
the source code. That's why we should exclude this dll file from the
release `.\bin` package.

> Note: by adding this dll to the `.\bin` folder of the release package,
we can again save an agent **AND** the default steps of the agent is now
also added when creating an 'Empty' one.
Seems like it has effect on this as well.

 #9204
 #9203