Update the Send Notification step to also update the record afterwards
Because of the newly added functionality of the retry mechanism, we
should update update the corresponding record accordingly to the
notification result.
The status of the notification result will result in either a 'Success',
'RetryableFail' or 'FatalFail'.

This commit also contains the necessary tests to verify if the related
record is set correctly for the status of the result of the
notification being sent.
 #9212