Adds the <PollingInterval/> to the 'settings.xml'
This commit contains the changes to add the a polling interval to the
'settings.xml' file; but also some major improvements on the `Config`
class to make sure that we don't get `NullReferenceException`'s on
missing tags.
I've tested this with all the tags gone/present; so it's a lot more
robust now.

Idea: we should maybe make some xsd files for the settings also?

 #9251