Update the Conformance Test 'Submit' transformer
The 'Submit' transformer should only handle `AS4Message`'s that are
'Submit' representations. This was not the case, which resulted in a
`MessagingContext` containing only a `ReceivedMessage`.

That's the reason why the `NullReferenceExcpetion` was thrown in the
first step: because the `AS4Message` was `null`.

This is now changed and tested with a message that 'bypasses' the AS4
Interceptor (like when the interceptor is down).