Reorganize documentation about Custom Connectors
The file `custom-connector.md` contained all the different types of
custom connectors, this is now changed so this file is the 'entry-point'
but all the different kind of connectors are split in different files.

> Note: since the `IDeliverSender` and `IAttachmentUploader` are now
having a different return type/responsibility. This is also added in the
documentation (they now return a `Send/UploadResult` to mark the
message/attachment possible for retry)
 #8990