BUG: update serialization of 'PartyId'
Same as the `Service` has the `PartyId` a optional type. This type
wasn't correctly serialized due to the mapping which had an ignored
validation of the members which resulted in overriding of already set
properties in the serialized models.

This commit fixed that issue and updated the validation of the deliver
model because this also can contain a non-existing type for the
`PartyId`.