Introduce PiggyBack functionality on the 'Pull Send' agent
This commit contains the changes and tests to make sure that the 'Pull
Send' agent can handle piggy-backed `PullRequest`'s with `Receipt`'s and
`Error`'s.
It includes 4 component tests to both test for an available
and not available `UserMessage` to be pulled, and piggybacked with a
`Receipt` and `Error`.

 #9578
 #8941