Add Sample Test with Windows Service as the Sender MSH
All the Sample Tests where made with the 'Console' instance as the
'Sender MSH' and the 'Windows Service' as the 'Receiver MSH'.
With this approach, a possible problem with the sending of messages by
the Windows Service wouldn't be detected. That's why this commit adds a
single Sample Test that sends the third sample (full blown sample) from
the Windows Service to the Console MSH to verify that the Windows
Service also can send out AS4Messages without problems.

 #9588