Add 'Origin' and 'Length' to the 'ReceivedMessage'
For better understanding from where exactly the incoming stream if
originated and how it internally looks. We should extend the
`ReceivedMessage` class.

The `Length` is also added so we could also check on this.
In the context of security, it's a good approach to take these into account:
- Origin of data (is now part of the `ReceivedMessage`)
- Size of data (is now part of the `ReceivedMessage`)
- Lexical Content of data (validate against a XSD, `ContentType` is part
of the `ReceivedMessage`))
- Syntax of data (during SOAP/MIME deserialization)
- Semantics of data (further during the steps, domain validation, some
during the deserialization)

 #6797