Insert AS4 Error for 'DeadLettered' message during Send-operation
During the migration of the ReceptionAwarenessAgent to the RetryAgent, the fact
that we insert a AS4 Error when the retry operation is completed as
'DeadLettered' so we can notify on this error wasn't migrate with it.
This commit adds this insertion to the RetryAgent as a special-case for
the Send-operation.

 #9819