Replace In/OutException MessageBody with MessageLocation
Just like the In/OutMessages, we should improve the saving of
In/OutExceptions by using a location instead of the full body of the
message that caused the exception.

This commit also contains the introduction of the `ExceptionService`
which warps these two operations (save of record + save of body) into a
single operation that should be used (just like the savings of
In/OutMessages).

 #9909
 #9880
 #9660
 #9501