FIX: make the In/OutException ctors private
We have introduced Factory Methods to create In/OutExceptions either by
Ebms Message Id or by Message Location. This way we have stricten the
creation but several tests were still using the 'old' way (directly the
ctor of the In/OutException).

This commit replaces these calls with the corresponding Factory Methods
and makes the ctors of these entities private.