BUG: add missing operations and status to monitoring query
Both the querying of messages and exceptions was out-of-date due to the
additional operations and statusses that were introduced.

This commit contains the necessary changes to update these values plus
makes sure that we have a more maintainable solution when we decide in
the future to add yet some more.

 #10235