BUG: monitoring querying with PMode was selecting content io pmode id
When querying for records with a specific PMode id, the query failed to
give the expected results due to the query that was setup was using the
entire PMode content to match instead of the PMode id.

This commit changes this to the PMode id, resulting in the expected
query results for the monitoring of messages.

 #10236