BUG: saving Sending PModes with/without Dynamic Discovery
Because we introduced the Dynamic Discovery option in the Sending PMode,
that means that we have or haven't a 'Tls Configuration'.
This caused problems because we tried to deserialize the Tls
configuration from a JObject > TlsConfiguration and this destination
type has a property of type `X509Type`.

This results in a `ArgumentNullException` because the default JSON
deserializer uses the `Enum.Parse` method but doesn't get any values if
we configure the Sending PMode as a Dynamic Discovered one.

This is now fixed by deserializing the properties of the Tls
configuration manually and in a more safe fashion. Meaning we have more
control on whether we have a value for this model or not.

 #9990