BUG: SMP Configuration couldn't be saved correctly with Encryption Certificate
The BUG about the saval of the 128 encryption keysize was with the older
version of the SMP Configuration window screen as this was discoverd via
an user of the component.

This can't be reproduced with the new SMP Configuration detail window
but there was an issue with the savings of encryption certificates and
the serialize/deserialize.

This was not done correctly and is changed with this commit to use both
the BASE64 encoding for serialize/deserialize of the certificate.

 #9989