TEST: refactoring the negative receive scenarios in integration tests
Place all the same negative scenarios for integration tests in the same
test file instead of placing them all across in different folders.
Since most of the integration tests are just placing files, it is not
much added-value for them to place them in separate folders and files
since it merely contains a single method per integration test.