REFACTER: remove unnecessary complex Deliver mapping
Remove the unnecessary complex mapping during the creation of the
DeliverMessage. Because we both need the SendingPMode and a UserMessage,
their exists also actions after the mapping, this not only makes the
mapping more complex; but also decreases the maintainability.

We have only one place where we map from 'Core' models to 'Common' (bad
name) models and that's in the Deliver operation.

A possible idea of mine would be in time to completely remove the
AutoMapper reference since in a lot of places this only increases
complexity.