STYLE: proposal for cleaner AS4 signal message creations
AS4 signal messages are always created based on an user message; but we
haven't yet made this very clear in the design. Especially if we want to
expose the core AS4 functionality it seems reasonable to add some
factory methods on the receipt and error model to create in a simple
fashion references signal messages.