FIX: also insert the ReceivingPMode for response/sync SignalMessages
Before we only had a ReceivingPMode when we respond the SignalMessage in
a Callback manner; but since we included the required values from the
'ResponsePMode' into the ReceivingPMode we also have a ReceivingPMode
for SignalMessages that we respond in a synchronous manner.

Therefore it seems logical to also include the ReceivingPMode for these
messages in the datastore.