FEAT: add Oasis fallback by calling SMP REST binding wh Document Identifer & Scheme
If we call the SMP REST binding that we have retrieved from the DNS
NAPTR response without the Document Identifier and the Document Scheme;
we get a SMP response with the complete URI (Document Identifier and
Scheme filled in). This acts as a fallback and can be used within our
dynamic discovery profile.

 #10649