BUG: override ToParty in SubmitMessage after Dynamic Discovery
When we use Oasis Dynamic Discovery in a Submit scenario, we need a
ToParty element with both a Id and Type. When the SMP response contains
a <Certificate/> element, the ToParty element in the Sending PMode
that's being dynamically populated will be overwritten.

This overwritten ToParty should be the ToParty of the UserMessage but
since we Submit information overrules PMode information we can't do
this.

This commit makes sure that the `DynamicDiscoveryStep` will remove the
ToParty from the SubmitMessage when the `IDynamicDiscoveryProfile`
implementation has overwritten the ToParty. Making the overwritten
ToParty in the Sending PMode the final ToParty in the UserMessage.