FEAT: add 'Configure' & 'Open Portal' to windows service system tray
We need to configure first the portal url so we can be sure that we
launch the right url. This is needed because the portal can be
configured in many different ways and hosted on many different
locations; meaning we have lots of places that could go wrong.

This is all unnecessary if we first ask the operator on what url the
portal is hosted.