FEAT: monitoring filter query with unique PMode names
Because we don't have any difference in the datastore for sending and
receiving pmodes, we have also no difference in the query filter in the
monitoring page.

This means that we have duplicate entries for pmodes with the same name
but since we only use the name itself; it doesn't matter from which
pmode we use it.

This commit filters out the unique pmode names when the sending and
receiving pmodes are concatenated together.

 #10234