STYLE: reduce/simplify logging during the step execution
We have a lot of logs (which is good) but a lot of them are actually
trace logs and don't add any value while just sending/receiving some
messages.

This commit both reduces/simplifies the logging by moving some debug
logs to the trace log, and adding some info logs to make the user more
aware of what we send and receive.