BUG: encrypted & signed message with IsserSerial or KeyIdentifier adapts wrong KeyInfo
Because of the wide XPath selection in both the
`KeyIdentifierSecurityTokenReference` and
`IssuerSerialSecurityTokenReference` the `<KeyInfo/>` element of the
encryption information was adapted (adding a `Id` attribute) and not the
`<KeyInfo/>` element in the `<Signature/>`.

This resulted in an decryption failure on the receiving MSH.
This commit removes the addition of the `Id` XML attribute since the
ebMS XSD specifies this as 'optional'.