FEAT: stop execution error steps when '.AndStopExecution()' was called
Before, we didn't take into count the `CanProceed` property of the
`StepResult` when we go from the 'Normal Pipeline' to the 'Error
Pipeline'. Because of this, every failed step in the normal pipeline
would get send to the error pipeline.

This is almost always the case; but for example the `SendAS4MessageStep`
needed to throw an exception to stop the furhter execution while we
could make use of the `.AndStopExecution()` call.

And this functionality is exactly implemented in this commit.

 #10923