Merged PR 3838: DOC: explanation of the Windows Service MSI installation, starting & configuration
DOC: explanation of the Windows Service MSI installation, starting & configuration

The Windows Service is now installed via a MSI which is not completely
explained in the WIKI docs. This commit changes that together with an
walkthrough on how the system tray application that gets installed
together with the Windows Service works.

Cherry picked from !3836